refactor: Input 분리
